msm: kgsl: Remove POPP
POPP constantly attempts to lower the GPU's frequency behind the
governor's back in order to save power; however, the GPU governor in use
(msm-adreno-tz) is very good at determining the GPU's load and selecting
an appropriate frequency to run the GPU at.

POPP was created long ago, perhaps when msm-adreno-tz didn't exist or
didn't work so well, so it is clearly deprecated. Remove it.

Signed-off-by: Sultan Alsawaf <sultan@kerneltoast.com>
Change-Id: I70e4463ab9ea870c3bd007b7ee856822ea72c923
Signed-off-by: NotZeetaa <rodrigo2005contente@gmail.com>
Signed-off-by: itsurboimasarou <medicspam6@gmail.com>